HashedReleaseFixture uses OS agnostic paths
